+++ src/View/Listener/ManageCategoriesPanelListener.java
@@ -12,9 +12,9 @@ import java.awt.event.ActionListener;
 
 public class ManageCategoriesPanelListener implements ActionListener {
 
-    AppFrame appFrame;
-    JTable table;
-    CategoriesOperationDialog categoriesOperationDialog;
+    private AppFrame appFrame;
+    private JTable table;
+    private CategoriesOperationDialog categoriesOperationDialog;
 
     public final static String BTN_ADD_CATEGORIES = "btnAdd";
     public final static String BTN_EDIT_CATEGORIES = "btnEdit";
@@ -28,7 +28,7 @@ public class ManageCategoriesPanelListener implements ActionListener {
     @Override
     public void actionPerformed(ActionEvent e) {
         String cmd = e.getActionCommand();
-
+        String esit;
         switch (cmd){
             case BTN_ADD_CATEGORIES:
                 if (table.getSelectionModel().isSelectionEmpty()){
@@ -39,22 +39,32 @@ public class ManageCategoriesPanelListener implements ActionListener {
                 }
                 break;
             case BTN_EDIT_CATEGORIES:
-                ICategoria cat = CategoriaBusiness.getInstance().findByID(Integer.parseInt(table.getModel().getValueAt(table.getSelectedRow(), 0).toString()));
-                new CategoriesOperationDialog(appFrame, true, false, cat);
+                if (table.getSelectedRowCount()==1){
+                    ICategoria cat = CategoriaBusiness.getInstance().findByID(Integer.parseInt(table.getModel().getValueAt(table.getSelectedRow(), 0).toString()));
+                    new CategoriesOperationDialog(appFrame, true, false, cat);
+                }else {
+                    esit = "Devi selezionare un elemento per modificarlo";
+                    JOptionPane.showMessageDialog(appFrame, esit, "Errore", JOptionPane.ERROR_MESSAGE);
+                }
                 break;
             case BTN_DELETE_CATEGORIES:
-                String esit;
-                int idToDelete = Integer.parseInt(table.getModel().getValueAt(table.getSelectedRow(), 0).toString());
-                int i = CategoriaBusiness.getInstance().deleteByID(idToDelete);
-                if (i==1){
-                    esit = "Prodotto eliminato con successo!";
-                    JOptionPane.showMessageDialog(appFrame, esit, "Successo", JOptionPane.INFORMATION_MESSAGE);
-                    appFrame.setCurrentMainPanel(new ManageCategoriesPanel(appFrame));
-                } else{
-                    esit = "Errore durante l'eliminazione";
+                if (table.getSelectedRowCount()==1){
+                    int idToDelete = Integer.parseInt(table.getModel().getValueAt(table.getSelectedRow(), 0).toString());
+                    int i = CategoriaBusiness.getInstance().deleteByID(idToDelete);
+                    if (i==1){
+                        esit = "Prodotto eliminato con successo!";
+                        JOptionPane.showMessageDialog(appFrame, esit, "Successo", JOptionPane.INFORMATION_MESSAGE);
+                        appFrame.setCurrentMainPanel(new ManageCategoriesPanel(appFrame));
+                    } else{
+                        esit = "Errore durante l'eliminazione";
+                        JOptionPane.showMessageDialog(appFrame, esit, "Errore", JOptionPane.ERROR_MESSAGE);
+                        appFrame.setCurrentMainPanel(new ManageCategoriesPanel(appFrame));
+                    }
+                }else{
+                    esit = "Devi selezionare un elemento per cancellarlo";
                     JOptionPane.showMessageDialog(appFrame, esit, "Errore", JOptionPane.ERROR_MESSAGE);
-                    appFrame.setCurrentMainPanel(new ManageCategoriesPanel(appFrame));
                 }
+
                 break;
         }
     }
